stylesjs added to big components where styles were declared in the bottom.
if a small component with the makestyleS() declared in the jsx remains, thats not big deal.
