+++ client/src/Views/Cart/styles.js
@@ -1,0 +1,46 @@
+import {
+    makeStyles
+} from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+    root: {
+        borderRadius: theme.shape.borderRadius,
+    },
+    head: {
+        backgroundColor: theme.palette.primary.main,
+    },
+    title: {
+        color: theme.palette.secondary.main,
+        fontWeight: 'bold',
+    },
+    body: {
+        backgroundColor: theme.palette.secondary.main,
+    },
+    tableContainer: {
+        marginBottom: 15,
+    },
+    productName: {
+        marginLeft: 15,
+    },
+    divider: {
+        color: theme.palette.secondary.dark,
+    },
+    quantityInput: {
+        width: 75,
+        borderRadius: 5,
+        backgroundColor: theme.palette.common.white,
+    },
+    subtotal: {
+        margin: 25,
+    },
+    checkout: {
+        margin: 5,
+    },
+    modal: {
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+}));
+
+export default useStyles;
