+++ client/src/Views/Footer/Footer.jsx
@@ -1,5 +1,4 @@
 //Footer
-
 import { Box, Container, Grid, Link } from "@material-ui/core";
 import React from "react";
