+++ client/src/components/Catalogue/Catalogue.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector, useDispatch } from "react-redux";
 import { getProducts, updateSearching, updateCategory } from '../../actions/home/home_actions';
-import { fade, makeStyles } from '@material-ui/core/styles';
+import useStyles from './styles'
 import { Grid, Paper, InputBase, Typography } from '@material-ui/core';
 import SearchIcon from '@material-ui/icons/Search';
 import ProductCards from '../ProductCards/ProductCards.jsx'
@@ -65,81 +65,4 @@ export default function Catalogue() {
     )
 };
 
-// Styles
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  catalogueMainContainer: {
-      backgroundColor: theme.palette.secondary.main,
-      boxShadow: '1px 1px 15px -1px rgba(0,0,0,0.6)',
-      width: '100%',
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'center',
-      height: '100%',
-      padding: '15px',
-      borderRadius: '10px',
-  },
-  searchSortContainer: {
-    display: 'flex',  
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    width: '98%',
-    borderRadius: '10px',
-    padding: '5px',
-    marginBottom: '24px',
-  },
-  search: {
-    position: 'relative',
-    display: 'flex',
-    alignItems: 'center',
-    height: '100%',
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: theme.palette.common.white,
-    '&:hover': {
-      backgroundColor: fade(theme.palette.common.white),
-    },
-    marginLeft: 0,
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      marginLeft: theme.spacing(1),
-      width: 'auto',
-    },
-  },
-  searchIcon: {
-    padding: theme.spacing(0, 2),
-    height: '100%',
-    position: 'absolute',
-    pointerEvents: 'none',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  inputRoot: {
-    color: 'inherit',
-  },
-  inputInput: {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: '25ch',
-    },
-  },
-  gridContainer: {
-    display: 'flex',
-    justifyContent: 'space-evenly',
-  },
-  productCard: {
-    margin: '10px',
-  },
-  paginationContainer: {
-    display: 'flex',
-    justifyContent: 'center',
-    margin: '15px',
-    width: '100%'
-  }
-}));
+
