+++ client/src/components/HomeLeftBar/styles.js
@@ -1,0 +1,16 @@
+import {
+    makeStyles
+} from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+    filter: {
+        backgroundColor: theme.palette.primary.main,
+        paddingBottom: theme.spacing(40),
+        paddingTop: theme.spacing(3),
+    },
+    links: {
+        color: theme.palette.secondary.main,
+    },
+}));
+
+export default useStyles;
