+++ client/src/components/ProductDetailsCard/ProductDetailsCard.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from "react-redux";
-import { makeStyles } from '@material-ui/core/styles';
+import useStyles from './styles';
 //Import components
 import ProductDetailsTab from './ProductDetailsTab/ProductDetailsTab.jsx'
 import ProductDetailsPhotoSlider from './ProductDetailsPhotoSlider/ProductDetailsPhotoSlider.jsx'
@@ -196,46 +196,3 @@ export default function ProductDetailsCard({ product, scoreArray, setModalState 
     )
 }
 
-
-//Custom styles
-const useStyles = makeStyles((theme) => ({
-    root: {
-      width: 750,
-      height: 500,
-      backgroundColor: theme.palette.secondary.main,
-      borderRadius: theme.shape.borderRadius,
-    },
-    container: {
-        height: '100%',
-      },
-    section: {
-        height: '100%',
-        width: '100%'
-    },
-    scoreContainer: {
-        width: 120,
-        marginBottom: 6,
-        marginLeft: 65
-    },
-    card: {
-        height: '100%',
-    },
-    name: {
-        marginRight: 17,
-    },
-    quantityInput: {
-        width: 80,
-        borderRadius: 5,
-        backgroundColor: theme.palette.common.white,
-      },
-    closeButton: {
-        margin: 0,
-        padding: 0,
-      },
-    cartTotal: {
-        marginLeft: 25,
-      },
-    favButton: {
-        width: 155,
-      },
-    }));
